services/auth/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.6
neo4j==5.15.0
pydantic[email]==2.5.0

psycopg2-binary==2.9.9
